requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
scikit-learn==1.3.0
numpy==1.24.3
